separate into 3 propagators to enable future improvements
