+++ src/test/java/org/chocosolver/solver/constraints/ternary/ModXYZTest.java
@@ -1,0 +1,34 @@
+/*
+ * This file is part of choco-solver, http://choco-solver.org/
+ *
+ * Copyright (c) 2019, IMT Atlantique. All rights reserved.
+ *
+ * Licensed under the BSD 4-clause license.
+ *
+ * See LICENSE file in the project root for full license information.
+ */
+package org.chocosolver.solver.constraints.ternary;
+
+import org.chocosolver.solver.Model;
+import org.chocosolver.solver.constraints.Constraint;
+import org.chocosolver.solver.variables.IntVar;
+
+/**
+ * <br/>
+ *
+ * @author Charles Prud'homme
+ * @since 16/07/12
+ */
+public class ModXYZTest extends AbstractTernaryTest {
+
+	@Override
+	protected int validTuple(int vx, int vy, int vz) {
+		return (vy != 0 && vz == vx - vy * (vx / vy)) ? 1 : 0;
+	}
+
+	@Override
+	protected Constraint make(IntVar[] vars, Model s) {
+		return s.mod(vars[0], vars[1], vars[2]);
+	}
+
+}
